Move code and start on twitter
Moved YouTube code from index to separate file
Added some code to start on Twitter feed, todo:
1. Figure out what the tweet object is in twit module
2. Add commands to follow and unfollow users
