+++ config.example.json
@@ -27,6 +27,12 @@
     "callbackUrl": "",
     "hubUrl": ""
   },
+  "twitter": {
+    "consumer_key": "",
+    "consumer_secret": "",
+    "app_only_auth": true,
+    "strictSSL": true
+  },
   "YtApiKey": "",
   "environment": "production",
   "logLevel": "info"
