+++ package.json
@@ -12,7 +12,7 @@
     "fluent-ffmpeg": "^2.1.2",
     "googleapis": "^58.0.0",
     "mongoose": "^5.9.27",
-    "node-fetch": "^2.6.0",
+    "node-fetch": "^2.6.1",
     "node-id3": "^0.1.18",
     "node-schedule": "^1.3.2",
     "pubsubhubbub": "^1.0.1",
@@ -20,6 +20,7 @@
     "query-string": "^6.13.1",
     "raw-body": "^2.4.1",
     "shortid": "^2.2.15",
+    "twit": "^2.2.11",
     "winston": "^3.3.3",
     "xml2js": "^0.4.23"
   },
